refacturización a todo el código y nueva interfaz, se agrega evento drop para el documento
