+++ index.html
@@ -5,22 +5,28 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./css/style.css" />
+    <link
+            rel="stylesheet"
+            href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css"
+        />
     <script src="./main.js" type="module" defer></script>
   </head>
   <body>
+    
     <div class="menu"></div>
 
     <div class="container">
       <div id="alert" class="alert-copy"></div>
-        <textarea id="text-svg" class="box text-svg"></textarea>
-        <div id="btn-procces" class="btn-copy"><h1>Procces</h1></div>
+      
       <div class="container-2">
-        <div id="svg" class="box svg"></div>
+        
+        <div id="listado" class="box listado" data-show></div>
+        <div id="svg" class="box svg" data-show></div>
         <!-- <ol id="listado" class="listado"></ol> -->
-        <div id="listado" class="listado"></div>
+        
       </div>
 
-      <div id="json" class="box json"></div>
+      <div id="json" class="box json" data-show></div>
       <textarea id="copy-json" class="copy-json"></textarea>
       <div class="btn-area">
         <div id="btn-copy" class="btn-copy"><h1>Copy</h1></div>
